+++ src/drvi/__init__.py
@@ -1,6 +1,6 @@
 from importlib.metadata import version
 
-from . import model, utils
+from . import internal, model, utils
 
 __version__ = version("drvi-py")
 
@@ -8,4 +8,5 @@ __all__ = [
     "__version__",
     "model",
     "utils",
+    "internal",
 ]