+++ src/drvi/internal/__init__.py
@@ -1,0 +1,4 @@
+from drvi.internal import drvi
+from drvi.internal.drvi import DRVI
+
+__all__ = ["DRVI", "drvi"]